Consolidated services and container.
