+++ pyshelf/search/container.py
@@ -1,12 +1,15 @@
-from elasticsearch import Elasticsearch
 from elasticsearch_dsl.connections import connections
+from pyshelf.search.update_manager import UpdateManager
+from pyshelf.search.manager import Manager as SearchManager
 
 
 class Container(object):
-    def __init__(self, logger, config):
+    def __init__(self, config, logger):
         self.logger = logger
         self.config = config
         self._hosts = None
+        self._update_manager = None
+        self._search_manager = None
         self.default_elastic_connection()
 
     @property
@@ -18,3 +21,17 @@ class Container(object):
 
     def default_elastic_connection(self):
         connections.create_connection("default", hosts=self.hosts)
+
+    @property
+    def update_manager(self):
+        if not self._update_manager:
+            self._update_manager = UpdateManager(self.logger)
+
+        return self._update_manager
+
+    @property
+    def search_manager(self):
+        if not self._search_manager:
+            self._search_manager = SearchManager(self)
+
+        return self._search_manager
