--- pyshelf/search/services.py
@@ -1,25 +1,0 @@
-from pyshelf.search.manager import Manager
-from pyshelf.search.container import Container as SearchContainer
-from pyshelf.search.update_manager import UpdateManager
-
-
-class Services(object):
-    def __init__(self, logger, config):
-        self.search_container = SearchContainer(logger, config)
-        self._update_manager = None
-        self._search_manager = None
-        self._search_parser = None
-
-    @property
-    def update_manager(self):
-        if not self._update_manager:
-            self._update_manager = UpdateManager(self.search_container.logger)
-
-        return self._update_manager
-
-    @property
-    def search_manager(self):
-        if not self._search_manager:
-            self._search_manager = Manager(self.search_container)
-
-        return self._update_manager
